requirements.txt
openai_whisper==20240930
pytest==8.3.5
python-dotenv==1.1.0
whisper==1.1.10
